Key pair refactoring
