+++ doc/operators/recovery.rst
@@ -31,9 +31,9 @@ To initiate the first phase of the recovery procedure, one or several nodes shou
     recover
     --network-cert-file /path/to/network_certificate
 
-Each node will then immediately restore the public entries of its ledger (``--ledger-dir``). Because deserialising the public entries present in the ledger may take some time, operators can query the progress of the public recovery by calling ``node/state`` which returns the version of the last signed recovered ledger entry. Once the public ledger is fully recovered, the recovered node automatically becomes part of the public network, allowing other nodes to join the network.
+Each node will then immediately restore the public entries of its ledger (``--ledger-dir``). Because deserialising the public entries present in the ledger may take some time, operators can query the progress of the public recovery by calling ``/node/state`` which returns the version of the last signed recovered ledger entry. Once the public ledger is fully recovered, the recovered node automatically becomes part of the public network, allowing other nodes to join the network.
 
-.. note:: If more than one node were started in ``recover`` mode, the node with the highest signed index (as per the response to the ``node/state`` RPC) should be preferred to start the new network. Other nodes should be shutdown and new nodes restarted with the ``join`` option.
+.. note:: If more than one node were started in ``recover`` mode, the node with the highest signed sequence number (as per the response to the ``/node/state`` RPC) should be preferred to start the new network. Other nodes should be shutdown and new nodes restarted with the ``join`` option.
 
 Similarly to the normal join protocol (see :ref:`operators/start_network:Adding a New Node to the Network`), other nodes are then able to join the network.
 
@@ -48,10 +48,10 @@ Similarly to the normal join protocol (see :ref:`operators/start_network:Adding 
         Node 2-->>Operators: Network Certificate
         Note over Node 2: Reading Public Ledger...
 
-        Operators->>+Node 2: node/state
+        Operators->>+Node 2: /node/state
         Node 2-->>Operators: {"last_signed_seqno": 50, "state": "readingPublicLedger"}
         Note over Node 2: Finished Reading Public Ledger, now Part of Public Network
-        Operators->>Node 2: node/state
+        Operators->>Node 2: /node/state
         Node 2-->>Operators: {"last_signed_seqno": 243, "state": "partOfPublicNetwork"}
 
         Note over Operators, Node 2: Operators select Node 2 to start the new network
