+++ getting_started/setup_vm/roles/ccf_build/vars/common.yml
@@ -27,5 +27,3 @@ debs:
 mbedtls_ver: "2.16.6"
 mbedtls_dir: "mbedtls-{{ mbedtls_ver }}"
 mbedtls_src: "{{ mbedtls_dir }}.tar.gz"
-step_cli_ver: "0.15.2"
-step_cli: "https://github.com/smallstep/cli/releases/download/v{{ step_cli_ver }}/step-cli_{{ step_cli_ver }}_amd64.deb"
